requirements.txt
crewai
litellm
google-genai
numpy
soundfile
streamlit
python-dotenv
